fastapi-backend/requirements.txt
fastapi
uvicorn
opencv-python
aiofiles
python-multipart
pyjwt
passlib